+++ src/components/Backgrounds/Spring/SpringStyle.jsx
@@ -1,9 +1,8 @@
 import styled, { keyframes } from 'styled-components'
-import { theme } from '../../utils/style/colors'
-import clouds from '../../assets/images/home/clouds.png'
-import branchBelow from '../../assets/images/home/branchBelow.png'
-import branchAbove from '../../assets/images/home/branchAbove.png'
-import mascotto from '../../assets/images/home/mascotto.png'
+import clouds from '../../../assets/images/home/clouds.png'
+import branchBelow from '../../../assets/images/home/branchBelow.png'
+import branchAbove from '../../../assets/images/home/branchAbove.png'
+import mascotto from '../../../assets/images/home/mascotto.png'
 
 const cloudAnimation = keyframes`
     0% { background-position: 1000px}
@@ -18,11 +17,11 @@ const Clouds = styled.div`
 `
 
 const HomePageBanner = styled.div`
-    background: 100% ${theme.home_page_bg_color};
+    background: 100% ${(props) => props.theme.home_page_bg_color};
     background: linear-gradient(
         0deg,
-        ${theme.home_gradient} 0%,
-        ${theme.home_page_bg_color} 62%
+        ${(props) => props.theme.home_gradient} 0%,
+        ${(props) => props.theme.home_page_bg_color} 62%
     );
     height: 100vh;
     width: auto;
@@ -42,4 +41,4 @@ const HomePageTreeAbove = styled.div`
     width: auto;
 `
 
-export { HomePageBanner, HomePageTreeAbove, HomePageTreeBelow, Clouds }
+export { Clouds, HomePageBanner, HomePageTreeAbove, HomePageTreeBelow }
